chat_pdf_app/requirements.txt
langchain==0.3.13
langchain-community==0.3.13
langgraph==0.2.62
langchain-ollama==0.2.3
streamlit==1.41.1
faiss-cpu==1.9.0.post1
PyMuPDF==1.25.2
tiktoken==0.8.0
numpy==1.26.4
python-dotenv==1.0.1
langchain_openai==0.3.13
langsmith
